requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
mysql-connector-python
pydantic-settings